refactor(agent): merge main and split Agent.swift beside AgentTurnKernel
Bring this branch onto current main so the split no longer rewrites the
Aug 19 god file. Keep AgentTurnKernel, AgentTurnDependencies, and
AgentRunEnvironment. Drop the shallow ToolLoopEngine wrapper.

The loop stays on Agent as Agent+ToolLoop, taking the once-per-turn
AgentTurnDependencies snapshot and asking AgentTurnKernel for decisions.
ConversationMessage stays private in the loop file; toolRegistry stays
private on Agent. Catalog source-file count: 186 → 193.
