+++ Sources/Swarm/Agents/Agent+Builder.swift
@@ -1,0 +1,432 @@
+// Agent+Builder.swift
+// Swarm Framework
+//
+// Extracted from Agent.swift. Effects stay on Agent behind AgentTurnKernel;
+// collaborators come from the once-per-turn AgentTurnDependencies snapshot.
+
+import Foundation
+
+// MARK: Agent.Builder
+
+public extension Agent {
+    /// Builder for creating Agent instances with a fluent API.
+    ///
+    /// Uses value semantics (struct) for Swift 6 concurrency safety.
+    ///
+    /// Example:
+    /// ```swift
+    /// let agent = Agent.Builder()
+    ///     .tools([WeatherTool(), CalculatorTool()])
+    ///     .instructions("You are a helpful assistant.")
+    ///     .configuration(.default.maxIterations(5))
+    ///     .build()
+    /// ```
+    struct Builder: Sendable {
+        // MARK: Public
+
+        // MARK: - Initialization
+
+        /// Creates a new builder.
+        public init() {}
+
+        // MARK: - Builder Methods
+
+        /// Sets the tools.
+        /// - Parameter tools: The tools to use.
+        /// - Returns: A new builder with the tools set.
+        @discardableResult
+        @available(*, deprecated, message: "Use tools(_:) with typed Tool values or Agent.withTools(@ToolBuilder:) for canonical typed tools.")
+        public func tools(_ tools: [any AnyJSONTool]) -> Builder {
+            var copy = self
+            copy._tools = tools
+            return copy
+        }
+
+        /// Sets the tools from typed tool instances.
+        /// - Parameter tools: The typed tools to use.
+        /// - Returns: A new builder with the tools set.
+        @discardableResult
+        public func tools(_ tools: [some Tool]) -> Builder {
+            var copy = self
+            copy._tools = tools.map { AnyJSONToolAdapter($0) }
+            return copy
+        }
+
+        /// Adds a tool (concrete type preferred; Swift resolves `some` before opening `any`).
+        /// - Parameter tool: The tool to add.
+        /// - Returns: A new builder with the tool added.
+        @discardableResult
+        @available(*, deprecated, message: "Use addTool(_:) with a typed Tool, or wrap raw tools in a clearly marked advanced adapter.")
+        public func addTool(_ tool: some AnyJSONTool) -> Builder {
+            var copy = self
+            copy._tools.append(tool)
+            return copy
+        }
+
+        /// Adds a tool from an existential (use when the concrete type is not available at the call site).
+        /// - Parameter tool: The tool to add.
+        /// - Returns: A new builder with the tool added.
+        @discardableResult
+        @available(*, deprecated, message: "Use addTool(_:) with a typed Tool, or wrap raw tools in a clearly marked advanced adapter.")
+        public func addTool(_ tool: any AnyJSONTool) -> Builder {
+            var copy = self
+            copy._tools.append(tool)
+            return copy
+        }
+
+        /// Adds a typed tool.
+        /// - Parameter tool: The typed tool to add.
+        /// - Returns: A new builder with the tool added.
+        @discardableResult
+        public func addTool(_ tool: some Tool) -> Builder {
+            var copy = self
+            copy._tools.append(AnyJSONToolAdapter(tool))
+            return copy
+        }
+
+        /// Adds built-in tools.
+        /// - Returns: A new builder with built-in tools added.
+        @discardableResult
+        public func withBuiltInTools() -> Builder {
+            var copy = self
+            copy._tools.append(contentsOf: BuiltInTools.all)
+            return copy
+        }
+
+        /// Sets the instructions.
+        /// - Parameter instructions: The system instructions.
+        /// - Returns: A new builder with the instructions set.
+        @discardableResult
+        public func instructions(_ instructions: String) -> Builder {
+            var copy = self
+            copy._instructions = instructions
+            return copy
+        }
+
+        /// Sets the configuration.
+        /// - Parameter configuration: The agent configuration.
+        /// - Returns: A new builder with the configuration set.
+        @discardableResult
+        public func configuration(_ configuration: AgentConfiguration) -> Builder {
+            var copy = self
+            copy._configuration = configuration
+            return copy
+        }
+
+        /// Sets the memory system.
+        /// - Parameter memory: The memory to use.
+        /// - Returns: A new builder with the memory set.
+        @discardableResult
+        public func memory(_ memory: any Memory) -> Builder {
+            var copy = self
+            copy._memory = memory
+            return copy
+        }
+
+        /// Sets the inference provider.
+        /// - Parameter provider: The provider to use.
+        /// - Returns: A new builder with the provider set.
+        @discardableResult
+        public func inferenceProvider(_ provider: any InferenceProvider) -> Builder {
+            var copy = self
+            copy._inferenceProvider = provider
+            return copy
+        }
+
+        /// Sets the tracer for observability.
+        /// - Parameter tracer: The tracer to use.
+        /// - Returns: A new builder with the tracer set.
+        @discardableResult
+        public func tracer(_ tracer: any Tracer) -> Builder {
+            var copy = self
+            copy._tracer = tracer
+            return copy
+        }
+
+        /// Sets the input guardrails.
+        /// - Parameter guardrails: The input guardrails to use.
+        /// - Returns: A new builder with the guardrails set.
+        @discardableResult
+        public func inputGuardrails(_ guardrails: [any InputGuardrail]) -> Builder {
+            var copy = self
+            copy._inputGuardrails = guardrails
+            return copy
+        }
+
+        /// Adds an input guardrail.
+        /// - Parameter guardrail: The guardrail to add.
+        /// - Returns: A new builder with the guardrail added.
+        @discardableResult
+        public func addInputGuardrail(_ guardrail: any InputGuardrail) -> Builder {
+            var copy = self
+            copy._inputGuardrails.append(guardrail)
+            return copy
+        }
+
+        /// Sets the output guardrails.
+        /// - Parameter guardrails: The output guardrails to use.
+        /// - Returns: A new builder with the guardrails set.
+        @discardableResult
+        public func outputGuardrails(_ guardrails: [any OutputGuardrail]) -> Builder {
+            var copy = self
+            copy._outputGuardrails = guardrails
+            return copy
+        }
+
+        /// Adds an output guardrail.
+        /// - Parameter guardrail: The guardrail to add.
+        /// - Returns: A new builder with the guardrail added.
+        @discardableResult
+        public func addOutputGuardrail(_ guardrail: any OutputGuardrail) -> Builder {
+            var copy = self
+            copy._outputGuardrails.append(guardrail)
+            return copy
+        }
+
+        /// Sets the guardrail runner configuration.
+        /// - Parameter configuration: The guardrail runner configuration.
+        /// - Returns: A new builder with the updated configuration.
+        @discardableResult
+        public func guardrailRunnerConfiguration(_ configuration: GuardrailRunnerConfiguration) -> Builder {
+            var copy = self
+            copy._guardrailRunnerConfiguration = configuration
+            return copy
+        }
+
+        /// Sets the handoff configurations.
+        /// - Parameter handoffs: The handoff configurations to use.
+        /// - Returns: A new builder with the updated handoffs.
+        @discardableResult
+        public func handoffs(_ handoffs: [AnyHandoffConfiguration]) -> Builder {
+            var copy = self
+            copy._handoffs = handoffs
+            return copy
+        }
+
+        /// Adds a handoff configuration.
+        /// - Parameter handoff: The handoff configuration to add.
+        /// - Returns: A new builder with the handoff added.
+        @discardableResult
+        public func addHandoff(_ handoff: AnyHandoffConfiguration) -> Builder {
+            var copy = self
+            copy._handoffs.append(handoff)
+            return copy
+        }
+
+        /// Adds a handoff target using typed options.
+        ///
+        /// This is the canonical front-facing handoff API.
+        ///
+        /// - Parameters:
+        ///   - target: The target agent.
+        ///   - configure: Optional typed options transformer.
+        /// - Returns: A new builder with the handoff added.
+        @discardableResult
+        public func handoff<Target: AgentRuntime>(
+            to target: Target,
+            configure: (HandoffOptions<Target>) -> HandoffOptions<Target> = { $0 }
+        ) -> Builder {
+            var copy = self
+            let options = configure(HandoffOptions())
+            copy._handoffs.append(options.erasedConfiguration(for: target))
+            return copy
+        }
+
+        /// Adds multiple handoff targets using Swift parameter packs.
+        ///
+        /// Example:
+        /// ```swift
+        /// let agent = try Agent.Builder()
+        ///     .handoffs(billingAgent, supportAgent, salesAgent)
+        ///     .build()
+        /// ```
+        @discardableResult
+        public func handoffs<each Target: AgentRuntime>(_ targets: repeat each Target) -> Builder {
+            var copy = self
+            repeat copy._handoffs.append(AnyHandoffConfiguration(targetAgent: each targets))
+            return copy
+        }
+
+        /// Builds the agent.
+        /// - Returns: A new Agent instance.
+        /// - Throws: `ToolRegistryError.duplicateToolName` if duplicate tool names are provided.
+        public func build() throws -> Agent {
+            try Agent(
+                tools: _tools,
+                instructions: _instructions,
+                configuration: _configuration,
+                memory: _memory,
+                inferenceProvider: _inferenceProvider,
+                tracer: _tracer,
+                inputGuardrails: _inputGuardrails,
+                outputGuardrails: _outputGuardrails,
+                guardrailRunnerConfiguration: _guardrailRunnerConfiguration,
+                handoffs: _handoffs
+            )
+        }
+
+        // MARK: Private
+
+        private var _tools: [any AnyJSONTool] = []
+        private var _instructions: String = ""
+        private var _configuration: AgentConfiguration = .default
+        private var _memory: (any Memory)?
+        private var _inferenceProvider: (any InferenceProvider)?
+        private var _tracer: (any Tracer)?
+        private var _inputGuardrails: [any InputGuardrail] = []
+        private var _outputGuardrails: [any OutputGuardrail] = []
+        private var _guardrailRunnerConfiguration: GuardrailRunnerConfiguration = .default
+        private var _handoffs: [AnyHandoffConfiguration] = []
+    }
+}
+
+// MARK: - Convenience Initializers
+
+public extension Agent {
+    /// Creates a new Agent with a name as the first parameter.
+    ///
+    /// This convenience initializer mirrors the OpenAI Agent SDK pattern
+    /// where the agent name is a top-level parameter rather than nested
+    /// inside configuration.
+    ///
+    /// Example:
+    /// ```swift
+    /// let agent = Agent(name: "Triage", instructions: "Route requests", tools: [weatherTool])
+    /// ```
+    ///
+    /// - Parameters:
+    ///   - name: The display name of the agent.
+    ///   - instructions: System instructions defining agent behavior. Default: ""
+    ///   - tools: Tools available to the agent. Default: []
+    ///   - inferenceProvider: Optional custom inference provider. Default: nil
+    ///   - memory: Optional explicit memory override. Default: ContextCore+Wax `DefaultAgentMemory` when Integrations is enabled; otherwise `SlidingWindowMemory`
+    ///   - tracer: Optional tracer for observability. Default: nil
+    ///   - configuration: Additional agent configuration settings. Default: .default
+    ///   - inputGuardrails: Input validation guardrails. Default: []
+    ///   - outputGuardrails: Output validation guardrails. Default: []
+    ///   - guardrailRunnerConfiguration: Configuration for guardrail runner. Default: .default
+    ///   - handoffs: Handoff configurations for multi-agent orchestration. Default: []
+    /// - Throws: `ToolRegistryError.duplicateToolName` if duplicate tool names are provided.
+    init(
+        name: String,
+        instructions: String = "",
+        tools: [any AnyJSONTool] = [],
+        inferenceProvider: (any InferenceProvider)? = nil,
+        memory: (any Memory)? = nil,
+        tracer: (any Tracer)? = nil,
+        configuration: AgentConfiguration = .default,
+        inputGuardrails: [any InputGuardrail] = [],
+        outputGuardrails: [any OutputGuardrail] = [],
+        guardrailRunnerConfiguration: GuardrailRunnerConfiguration = .default,
+        handoffs: [AnyHandoffConfiguration] = []
+    ) throws {
+        // Merge the name into the configuration
+        var config = configuration
+        config.name = name
+        try self.init(
+            tools: tools,
+            instructions: instructions,
+            configuration: config,
+            memory: memory,
+            inferenceProvider: inferenceProvider,
+            tracer: tracer,
+            inputGuardrails: inputGuardrails,
+            outputGuardrails: outputGuardrails,
+            guardrailRunnerConfiguration: guardrailRunnerConfiguration,
+            handoffs: handoffs
+        )
+    }
+}
+
+// MARK: - Simplified Handoff Declaration
+
+public extension Agent {
+    /// Creates an Agent with agents directly as handoff targets.
+    ///
+    /// This convenience initializer eliminates the need to wrap each agent
+    /// in `AnyHandoffConfiguration`, inspired by the OpenAI SDK pattern
+    /// where you pass agents directly: `Agent(handoffs=[billing, support])`.
+    ///
+    /// Example:
+    /// ```swift
+    /// let triage = Agent(
+    ///     name: "Triage",
+    ///     instructions: "Route requests",
+    ///     handoffAgents: [billingAgent, supportAgent]
+    /// )
+    /// ```
+    ///
+    /// - Parameters:
+    ///   - name: The display name of the agent.
+    ///   - instructions: System instructions. Default: ""
+    ///   - tools: Tools available to the agent. Default: []
+    ///   - inferenceProvider: Optional inference provider. Default: nil
+    ///   - memory: Optional explicit memory override. Default: ContextCore+Wax `DefaultAgentMemory` when Integrations is enabled; otherwise `SlidingWindowMemory`
+    ///   - tracer: Optional tracer. Default: nil
+    ///   - configuration: Additional configuration. Default: .default
+    ///   - inputGuardrails: Input guardrails. Default: []
+    ///   - outputGuardrails: Output guardrails. Default: []
+    ///   - guardrailRunnerConfiguration: Guardrail runner config. Default: .default
+    ///   - handoffAgents: Agents to use as handoff targets.
+    /// - Throws: `ToolRegistryError.duplicateToolName` if duplicate tool names are provided.
+    init(
+        name: String,
+        instructions: String = "",
+        tools: [any AnyJSONTool] = [],
+        inferenceProvider: (any InferenceProvider)? = nil,
+        memory: (any Memory)? = nil,
+        tracer: (any Tracer)? = nil,
+        configuration: AgentConfiguration = .default,
+        inputGuardrails: [any InputGuardrail] = [],
+        outputGuardrails: [any OutputGuardrail] = [],
+        guardrailRunnerConfiguration: GuardrailRunnerConfiguration = .default,
+        handoffAgents: [any AgentRuntime]
+    ) throws {
+        let handoffs = handoffAgents.map { agent in
+            AnyHandoffConfiguration(targetAgent: agent)
+        }
+        try self.init(
+            name: name,
+            instructions: instructions,
+            tools: tools,
+            inferenceProvider: inferenceProvider,
+            memory: memory,
+            tracer: tracer,
+            configuration: configuration,
+            inputGuardrails: inputGuardrails,
+            outputGuardrails: outputGuardrails,
+            guardrailRunnerConfiguration: guardrailRunnerConfiguration,
+            handoffs: handoffs
+        )
+    }
+}
+
+// MARK: - V3 Canonical Init with Explicit Provider
+
+public extension Agent {
+    /// V3 convenience init with an explicit, non-optional inference provider.
+    ///
+    /// This overload avoids the optional wrapping when a provider is always known:
+    /// ```swift
+    /// let agent = try Agent("You are helpful.", provider: .foundationModels()) {
+    ///     WeatherTool()
+    /// }
+    /// ```
+    ///
+    /// - Parameters:
+    ///   - instructions: System instructions defining agent behavior.
+    ///   - provider: The inference provider to use.
+    ///   - tools: A `@ToolBuilder` closure producing the agent's tools. Default: empty.
+    /// - Throws: `ToolRegistryError.duplicateToolName` if duplicate tool names are provided.
+    init(
+        _ instructions: String,
+        provider: some InferenceProvider,
+        @ToolBuilder tools: () -> ToolCollection = { .empty }
+    ) throws {
+        try self.init(
+            tools: tools().storage,
+            instructions: instructions,
+            inferenceProvider: provider
+        )
+    }
+}
